+++ lab2/mean.cpp
@@ -2,7 +2,7 @@
 // Created by vlad on 21.02.2020.
 //
 
-int mean (int *a,  int n){
+float mean (int *a,  int n){
     int result = 0;
     for (int i = 0; i < n; i++)
         result += a[i];
